Update Consignment model with validation and cost calculation
Added validation for start and end dates, quantity field, and validation annotations in `ConsignmentRequest`. Modified `consignmentService` to apply the new cost calculation logic. Changed `costPerDay` to double in `CareType` and updated dependencies in `pom.xml`.
