+++ demo/src/main/java/com/example/demo/service/CertificateService.java
@@ -1,12 +1,16 @@
 package com.example.demo.service;
 
-import com.example.demo.entity.Certificate;
+import com.example.demo.entity.*;
 import com.example.demo.exception.NotFoundException;
+import com.example.demo.model.Request.EmailDetails;
 import com.example.demo.repository.CertificateRepository;
+import com.example.demo.repository.KoiRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.io.File;
+import java.util.Date;
 import java.util.List;
 
 @Service
@@ -15,8 +19,53 @@ public class CertificateService {
     private CertificateRepository certificateRepository;
     @Autowired
     private ModelMapper modelMapper;
+    @Autowired
+    private KoiRepository koiRepository;
+    @Autowired
+    private CertificatePdfGeneratorService certificatePdfGeneratorService;
+    @Autowired
+    private EmailService emailService;
+    private void createCertificatesAndSendEmail(Orders orders) {
+        for (OrderDetails orderDetail : orders.getOrderDetails()) {
+            Koi koi = orderDetail.getKoi();
 
-    // Tạo mới Certificate
+            // Create and save certificate for each Koi
+            Certificate certificate = new Certificate();
+            certificate.setKoi(koi);
+            certificate.setVariety(koi.getName());
+            certificate.setBreeder(koi.getVendor());
+            certificate.setBornIn(koi.getBornYear());
+            certificate.setSize(koi.getSize());
+            certificate.setImageUrl(koi.getImages());
+            certificate.setIssueDate(new Date());
+
+            certificateRepository.save(certificate);
+            koi.setCertificate(certificate);
+            koiRepository.save(koi);
+
+            // Send certificate email
+            sendCertificateEmail(orders.getCustomer(), certificate);
+        }
+    }
+    private void sendCertificateEmail(Account customer, Certificate certificate) {
+        try {
+            // Generate PDF from HTML
+            File pdfFile = certificatePdfGeneratorService.createCertificatePdf(certificate);
+
+            // Prepare email details
+            EmailDetails emailDetails = new EmailDetails();
+            emailDetails.setReceiver(customer);
+            emailDetails.setSubject("Your Koi Certificate");
+            emailDetails.setLink("http://koishop.site/");
+
+            // Send the email with the PDF attachment
+            emailService.sendEmailWithAttachment(emailDetails, pdfFile);
+        } catch (Exception e) {
+            e.printStackTrace();
+            // Handle exception (log or retry)
+        }
+    }
+
     // Lấy tất cả Certificates
     public List<Certificate> getAllCertificates() {
         return certificateRepository.findAll();
